+++ ivtest/vpi/nextsimtime_common.c
@@ -20,30 +20,57 @@
 # include  "vpi_user.h"
 # include  <assert.h>
 
-#if defined(TEST_SCALED) || defined(TEST_SUPPRESS)
-#if defined(TEST_SCALED)
+#if     defined(TEST_SIM_TIME)
+#define TIME_TYPE vpiSimTime
+#elif   defined(TEST_SCALED_TIME)
 #define TIME_TYPE vpiScaledRealTime
-#elif defined(TEST_SUPPRESS)
+#else
 #define TIME_TYPE vpiSuppressTime
 #endif
+
+#ifndef TEST_NULL_TIME
 static s_vpi_time cb_timerec = {TIME_TYPE, 0, 0, 0};
 #endif
 
-static PLI_INT32 dummy_cb(struct t_cb_data* cb) {
-      (void)cb; /* unused */
+static PLI_INT32 register_nextsimtime(struct t_cb_data* cb);
+
+
+static PLI_INT32 nextsimtime_cb(struct t_cb_data* cb) {
+      s_vpi_time timerec;
+#ifdef TEST_SCALED_TIME
+      timerec.type = vpiScaledRealTime;
+#else
+      timerec.type = vpiSimTime;
+#endif
+      vpi_get_time(NULL, &timerec);
+#ifdef TEST_SCALED_TIME
+      vpi_printf("nextsimtime: %f vpi_get_time: %f\n",
+            cb->time->real, timerec.real);
+#else
+      uint64_t nextsimtime = ((uint64_t)cb->time->high << 32) | cb->time->low;
+      uint64_t time = ((uint64_t)timerec.high << 32) | timerec.low;
+      vpi_printf("nextsimtime: %" PLI_UINT64_FMT " vpi_get_time: %" PLI_UINT64_FMT "\n",
+            nextsimtime, time);
+#endif
+      register_nextsimtime(NULL);
       return 0;
 }
 
 static PLI_INT32 register_nextsimtime(struct t_cb_data* cb) {
       (void)cb; /* unused */
       s_cb_data cb_data;
-#if defined(TEST_SCALED) || defined(TEST_SUPPRESS)
+#ifdef TEST_NULL_TIME
+      cb_data.time   = 0;
+#else
       cb_data.time   = &cb_timerec;
-#elif defined(TEST_NULL)
-      cb_data.time   = NULL;
+#endif
+#ifdef TEST_SCALED_TIME
+      cb_data.obj    = vpi_handle_by_name("main", NULL);
+#else
+      cb_data.obj    = 0;
 #endif
       cb_data.reason = cbNextSimTime;
-      cb_data.cb_rtn = dummy_cb;
+      cb_data.cb_rtn = nextsimtime_cb;
       vpiHandle hndl = NULL;
       assert((hndl = vpi_register_cb(&cb_data)) && vpi_free_object(hndl));
       return 0;
