+++ ivtest/vpi/nextsimtime_common.v
@@ -17,6 +17,8 @@
  *    Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  */
 
+`timescale 1s/1ms
+
 module main;
 
    initial begin
