+++ ivtest/vpi/nextsimtime_scaled_time.c
@@ -1,2 +1,2 @@
-#define TEST_SCALED
-#include "nextsimtime_errors.c"
+#define TEST_SCALED_TIME
+#include "nextsimtime_common.c"
