+++ ivtest/vpi/nextsimtime_sim_time.c
@@ -1,2 +1,2 @@
-#define TEST_SUPPRESS_TIME
+#define TEST_SIM_TIME
 #include "nextsimtime_common.c"
