+++ ivtest/vpi_gold/nextsimtime_sim_time.gold
@@ -1,0 +1,8 @@
+Compiling vpi/nextsimtime_sim_time.c...
+Making nextsimtime_sim_time.vpi from  nextsimtime_sim_time.o...
+time 0: 0
+nextsimtime: 1000 vpi_get_time: 1000
+time 1: 1
+nextsimtime: 4000 vpi_get_time: 4000
+time 4: 4
+nextsimtime: 5000 vpi_get_time: 5000
